+++ source/gui/settingsDialogs.py
@@ -2060,6 +2060,15 @@ class KeyboardSettingsPanel(SettingsPanel):
 
 		# Translators: This is the label for a checkbox in the
 		# keyboard settings panel.
+		maskPasswordsText = _("Mask passwords while typing")
+		self.maskPasswordsCheckBox = sHelper.addItem(
+			wx.CheckBox(self, label=maskPasswordsText),
+		)
+		self.bindHelpEvent("MaskPasswordsWhileTyping", self.maskPasswordsCheckBox)
+		self.maskPasswordsCheckBox.SetValue(config.conf["keyboard"]["maskPasswords"])
+
+		# Translators: This is the label for a checkbox in the
+		# keyboard settings panel.
 		handleInjectedKeysText = _("Handle keys from other &applications")
 		self.handleInjectedKeysCheckBox = sHelper.addItem(wx.CheckBox(self, label=handleInjectedKeysText))
 		self.bindHelpEvent("KeyboardSettingsHandleKeys", self.handleInjectedKeysCheckBox)
@@ -2108,6 +2117,7 @@ class KeyboardSettingsPanel(SettingsPanel):
 		config.conf["keyboard"]["beepForLowercaseWithCapslock"] = self.beepLowercaseCheckBox.IsChecked()
 		config.conf["keyboard"]["speakCommandKeys"] = self.commandKeysCheckBox.IsChecked()
 		config.conf["keyboard"]["alertForSpellingErrors"] = self.alertForSpellingErrorsCheckBox.IsChecked()
+		config.conf["keyboard"]["maskPasswords"] = self.maskPasswordsCheckBox.IsChecked()
 		config.conf["keyboard"]["handleInjectedKeys"] = self.handleInjectedKeysCheckBox.IsChecked()
 		config.conf["keyboard"]["multiPressTimeout"] = self.multiPressTimeoutEdit.GetValue()